requirements.txt
streamlit
textblob
openai
pandas
matplotlib
python-dotenv
